pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copilota"
version = "0.1.0"
description = "Asistente de código local con RAG, multi-lenguaje y vector search"
requires-python = ">=3.10"
dependencies = [
    "tree-sitter>=0.22",
    "tree-sitter-python>=0.21",
    "tree-sitter-javascript>=0.21",
    "tree-sitter-php>=0.22",
    "tree-sitter-go>=0.21",
    "tree-sitter-rust>=0.21",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pydantic>=2.6",
    "click>=8.1",
    "GitPython>=3.1",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
copilota = "copilota.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
